Join test from general and no-bid Renamed tests
